relevent changes

+++ GlobalDataUploder/leave/viewleave_leave.php
@@ -1,6 +1,17 @@
 <?php
+error_reporting(0);
 MysqlConnection::connect();
 $resource = MysqlConnection::fetchAll("tbl_addleave");
+if (isset($_POST["btnSearch"])) {
+    $leavetype = $_POST["leave_type"];
+    $nocount = $_POST["no_count"];
+
+    $sql_custom = "SELECT * FROM `tbl_addleave` "
+            . " WHERE `leave_type` LIKE '%$leavetype%' "
+            . " AND `no_count` LIKE '%$nocount%' ";
+
+    $resource = MysqlConnection::fetchCustom($sql_custom);
+}
 ?>
 <div class="container-fluid">
     <div class="row-fluid">
@@ -10,20 +21,29 @@ $resource = MysqlConnection::fetchAll("tbl_addleave");
                     <h5>SEARCH LEAVE </h5>
                 </div>
                 <div class="widget-content nopadding">
-                    <form class="form-horizontal" method="post" action="../OffsetuserServlet" name="basic_validate" id="basic_validate" novalidate="novalidate">
+                    <form class="form-horizontal" method="post" action="" name="basic_validate" id="basic_validate" novalidate="novalidate">
                         <div class="control-group">
-
                             <label class="control-label ">LEAVE TYPE :</label>
                             <div class="controls">
-                                <input type="text" name="leave_type"   autofocus="" maxlength="100" class="span11"    placeholder="USERNAME" />
-                            </div>
-                            <div class="controls">
-                                <input type="hidden" name="<%= IServletConstant.ACTION%>" value="<%= IServletConstant.ACTION_SEARCH%>" />
-                                <button type="submit" class="btn btn-success">Search</button>
-                                <button type="submit" class="btn btn-danger">Clear</button>
-                                  <a href="index.php?requestPage=addleave_leave"><button type="button" class="btn btn-info">Add</button></a>
+                                <input type="text" name="leave_type" value="<?php echo $leavetype ?>"  autofocus="" maxlength="100" class="span11"    placeholder="" />
                             </div>
                         </div>
+                        <div class="control-group">
+                            <label class="control-label ">NO OF COUNT :</label>
+                            <div class="controls">
+                                <input type="text" name="no_count" value="<?php echo $nocount ?>"  autofocus="" maxlength="100" class="span11"    placeholder="" />
+                            </div>
+                        </div>
+                        <div class="controls">
+                            <center>
+
+                                <button type="submit" name="btnSearch" class="btn btn-success">Search</button>
+                                <button type="submit" class="btn btn-danger">Clear</button>
+                                <a href="index.php?requestPage=addleave_leave"><button type="button" class="btn btn-info">Add</button></a>
+
+                            </center>
+                        </div>
+
                     </form>
                 </div>
             </div>
@@ -36,14 +56,14 @@ $resource = MysqlConnection::fetchAll("tbl_addleave");
         <div class="span12">
             <div class="widget-box">
                 <div class="widget-title"> <span class="icon"><i class="icon-th"></i></span>
-                    <h5>VIEW EMPLOYEE</h5>
+                    <h5>VIEW LEAVE</h5>
                 </div>
                 <div class="widget-content nopadding">
-                    <table class="table table-bordered data-table">
+                    <table class="table table-bordered data-table" style="font-size: 11px;">
                         <thead>
                             <tr>
-                                <th style="width: 1%">Edit</th>
-                                <th  style="width: 1%">Delete</th>
+                                <th style="width: 1%">#</th>
+                                <th  style="width: 1%">#</th>
                                 <th>Leave type</th>
                                 <th>Description</th>
                                 <th>No Of Count</th>
@@ -60,16 +80,16 @@ $resource = MysqlConnection::fetchAll("tbl_addleave");
                                 ?>
                                 <tr class="gradeX">
                                     <td>
-                                        <a href="index.php?requestPage=edit_leave&txtId=<?php echo $result["txtId"] ?>">
+                                        <a title="EDIT" href="index.php?requestPage=edit_leave&txtId=<?php echo $result["txtId"] ?>">
                                             <i class="icon-pencil"></i>
                                         </a>
                                     </td>
                                     <td>
-                                        <a href="index.php?requestPage=request_delete&tblname=tbl_addleave&pkvalue=<?php echo $result["txtId"] ?>&location=index.php?requestPage=viewleave_leave">
+                                        <a title="DELETE" onclick="return confirm('Are You Sure Want to delete this Record?');" href="index.php?requestPage=request_delete&tblname=tbl_addleave&pkvalue=<?php echo $result["txtId"] ?>&location=index.php?requestPage=viewleave_leave">
                                             <i class="icon-remove"></i>
                                         </a>                                  
                                     </td>
-                                   
+
                                     <td><?php echo $result["leave_type"] ?></td>
                                     <td><?php echo $result["description"] ?></td>
                                     <td><?php echo $result["no_count"] ?></td>
